refactor class user
