+++ app/src/main/java/com/picpay/desafio/android/repository/model/User.kt
@@ -1,0 +1,13 @@
+package com.picpay.desafio.android.repository.model
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class User(
+    @SerializedName("img") val img: String?,
+    @SerializedName("name") val name: String?,
+    @SerializedName("id") val id: Int?,
+    @SerializedName("username") val username: String?
+) : Parcelable
