+++ app/src/main/java/com/picpay/desafio/android/repository/remote/webclient/PicPayService.kt
@@ -1,5 +1,6 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.repository.remote.webclient
 
+import com.picpay.desafio.android.repository.model.User
 import retrofit2.Call
 import retrofit2.http.GET
